conform practices to C11 examples: sessionscontroller -> userscontroller
