--- app/helpers/sessions_helper.rb
@@ -1,2 +1,0 @@
-module SessionsHelper
-end
