--- test/controllers/sessions_controller_test.rb
@@ -1,4 +1,0 @@
-require "test_helper"
-
-describe SessionsController do
-end
